+++ src/IO/WKBByteOrder.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Brick\Geo\IO;
+
+enum WKBByteOrder: int
+{
+    case BIG_ENDIAN = 0;
+    case LITTLE_ENDIAN = 1;
+}