+++ tests/IO/EWKBWriterTest.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Brick\Geo\Tests\IO;
 
 use Brick\Geo\Exception\GeometryIOException;
-use Brick\Geo\IO\EWKBTools;
 use Brick\Geo\IO\EWKBWriter;
 use Brick\Geo\IO\EWKTReader;
+use Brick\Geo\IO\WKBByteOrder;
 use Brick\Geo\Point;
 
 /**
@@ -18,11 +18,11 @@ class EWKBWriterTest extends EWKBAbstractTest
     /**
      * @dataProvider providerWrite
      *
-     * @param string $ewkt      The EWKT to read.
-     * @param string $ewkb      The expected EWKB output, hex-encoded.
-     * @param int    $byteOrder The byte order to use.
+     * @param string       $ewkt      The EWKT to read.
+     * @param string       $ewkb      The expected EWKB output, hex-encoded.
+     * @param WKBByteOrder $byteOrder The byte order to use.
      */
-    public function testWrite(string $ewkt, string $ewkb, int $byteOrder) : void
+    public function testWrite(string $ewkt, string $ewkb, WKBByteOrder $byteOrder) : void
     {
         $writer = new EWKBWriter();
         $writer->setByteOrder($byteOrder);
@@ -38,19 +38,19 @@ class EWKBWriterTest extends EWKBAbstractTest
     public function providerWrite() : \Generator
     {
         foreach ($this->providerLittleEndianEWKB() as [$wkt, $ewkb]) {
-            yield [$wkt, $ewkb, EWKBTools::LITTLE_ENDIAN];
+            yield [$wkt, $ewkb, WKBByteOrder::LITTLE_ENDIAN];
         }
 
         foreach ($this->providerLittleEndianEWKB_SRID() as [$wkt, $ewkb]) {
-            yield [$wkt, $ewkb, EWKBTools::LITTLE_ENDIAN];
+            yield [$wkt, $ewkb, WKBByteOrder::LITTLE_ENDIAN];
         }
 
         foreach ($this->providerBigEndianEWKB() as [$wkt, $ewkb]) {
-            yield [$wkt, $ewkb, EWKBTools::BIG_ENDIAN];
+            yield [$wkt, $ewkb, WKBByteOrder::BIG_ENDIAN];
         }
 
         foreach ($this->providerBigEndianEWKB_SRID() as [$wkt, $ewkb]) {
-            yield [$wkt, $ewkb, EWKBTools::BIG_ENDIAN];
+            yield [$wkt, $ewkb, WKBByteOrder::BIG_ENDIAN];
         }
     }